Add coverage tooling to ensure tests have wrappers.
Since test cc_binary targets and sandbox_test generators are defined
separately, it is important to know whether or not there are generators
missing for a given target. E.g. someone may add foo_test, only to discover
that no foo_test targets are actually running, since an appropriate
sandbox_test macro had not also been added to the BUILD file.

This change adds a targets_check BUILD target that can consume a list of
targets (defined elsewhere) and ensure that a given set of targets *covers*
the defined set of targets. This error will manifest at build time.

The current "ignore" sets are not covered, and a follow-up will add those.

PiperOrigin-RevId: 369519281