Hold CtxIPCNamespace in kernel/ipc package.
CtxIPCNamespace is needed by mqfs package to be able to retreive an
IPCNamespace using ctx.Value. Unlike a map, ctx.Value only accepts the
original value used as a key when retreiving an object. This means that
we need access to the actual key, rather than a copy of it. Retreiving
the key from package kernel for use in mqfs would cause a circular
dependency, so we move to kernel/ipc.

Updates #136